moved RawResultKFoldValidator to cross_validation.py
